Add a new cors_rules variable to the aws-s3-static-website module to control the S3 bucket's CORS configuration.
The cors_rules variable contains a list of objects. Since the default value is an empty list ([]), users do not need to set this input variable to deploy the module. When they do use it, they must set allowed_methods and allowed_origins for each object in the list; the other attributes are optional. This matches the behavior of the aws_s3_bucket_cors_configuration resource used to configure CORS.